Self_Elna/requirements.txt
requests>=2.31.0
keyboard>=0.13.5
pyperclip>=1.8.2
mss>=9.0.1
Pillow>=10.0.0
screeninfo>=0.8.1
